+++ src/Service/PersonaServiceInterface.php
@@ -7,9 +7,24 @@ use Phamily\Framework\GenderAwareInterface;
 use Phamily\Framework\ValueObject\DateTimeInterface;
 use Phamily\Framework\KinshipAwareInterface;
 
+/**
+ *
+ * @author samizdam
+ *
+ */
 interface PersonaServiceInterface extends GenderAwareInterface, KinshipAwareInterface
 {
-    public function create($gender, array $names = [], PersonaInterface $father = null, PersonaInterface $mother = null, DateTimeInterface $dateOfBirth = null, DateTimeInterface $dateOfDeath = null);
+    /**
+     * Create new Persona with given attributes.
+     *
+     * @param string $gender
+     * @param array $names
+     * @param PersonaInterface $father
+     * @param PersonaInterface $mother
+     * @param DateTimeInterface $dateOfBirth
+     * @param DateTimeInterface $dateOfDeath
+     */
+    public function createPersona($gender, array $names = [], PersonaInterface $father = null, PersonaInterface $mother = null, DateTimeInterface $dateOfBirth = null, DateTimeInterface $dateOfDeath = null);
 
     public function delete(PersonaInterface &$persona);